Made X-UA-Compatible first <meta> tag in the header
The X-UA-Compatibility meta tag with `content="IE=edge"` asks IE11 to render using the most modern available settings. However, IE11 in compatibility mode only recognizes the `X-UA-Compatibility` tag when it is the **first** `<meta>` tag within the page header. This change should allow IE11 to work with MatterMost even when the domain on which MatterMost is hosted has been added to the browser's Compatibility View Settings site list.